Move bias function from Caffe function helpers to standard functions.
