+++ chainer/functions/__init__.py
@@ -45,6 +45,7 @@ from chainer.functions.loss import softmax_cross_entropy
 from chainer.functions.loss import vae  # NOQA
 from chainer.functions.math import basic_math  # NOQA
 from chainer.functions.math import batch_l2_norm_squared
+from chainer.functions.math import bias
 from chainer.functions.math import clip
 from chainer.functions.math import det
 from chainer.functions.math import exponential
@@ -178,6 +179,7 @@ BatchL2NormSquared = batch_l2_norm_squared.BatchL2NormSquared
 batch_l2_norm_squared = batch_l2_norm_squared.batch_l2_norm_squared
 BatchMatMul = matmul.BatchMatMul
 batch_matmul = matmul.batch_matmul
+bias = bias.bias
 Clip = clip.Clip
 clip = clip.clip
 Cos = trigonometric.Cos
