+++ docs/source/reference/functions.rst
@@ -239,6 +239,10 @@ batch_matmul
 ~~~~~~~~~~~~
 .. autofunction:: batch_matmul
 
+bias
+~~~~
+.. autofunction:: bias
+
 clip
 ~~~~
 .. autofunction:: clip
